Routing and test added
